+++ postpic/datareader/__init__.py
@@ -157,9 +157,17 @@ def chooseCode(code):
         Possible options are:
           - "DUMMY": dummy class creating fake data.
           - "EPOCH": .sdf files written by EPOCH1D, EPOCH2D or EPOCH3D.
-          - "openPMD": .h5 files written in openPMD Standard
-          - "piconGPU": same as "openPMD"
+            ( https://cfsa-pmw.warwick.ac.uk/EPOCH/epoch )
+          - "openPMD": .h5 files written in openPMD Standard.
+            ( https://github.com/openPMD/ )
+          - "fbpic": .h5 files written by fbpic ( https://github.com/fbpic/fbpic ).
+            This is the openPMD reader, but includes the conversion from azimuthal modes
+            to a cylindrical grid.
+          - "piconGPU": same as "openPMD".
+            This reader can be used for piconGPU written h5 files.
+            ( https://github.com/ComputationalRadiationPhysics/picongpu )
           - "VSIM": .hdf5 files written by VSim.
+            Currently not working. Needs updating.
     '''
     if code.lower() in ['epoch', 'epoch1d', 'epoch2d', 'epoch3d']:
         from .epochsdf import Sdfreader, Visitreader
@@ -169,6 +177,10 @@ def chooseCode(code):
         from .openPMDh5 import OpenPMDreader, FileSeries
         setdumpreadercls(OpenPMDreader)
         setsimreadercls(FileSeries)
+    elif code.lower() in ['fbpic']:
+        from .openPMDh5 import FbpicReader, FbpicFileSeries
+        setdumpreadercls(FbpicReader)
+        setsimreadercls(FbpicFileSeries)
     elif code.lower() in ['vsim']:
         raise Exception('VSim reader requires update due to the interface change in '
                         'https://github.com/skuschel/postpic/commit/'
@@ -181,4 +193,4 @@ def chooseCode(code):
         setdumpreadercls(Dummyreader)
         setsimreadercls(Dummysim)
     else:
-        raise TypeError('Code "' + str(code) + '" not recognized.')
+        raise TypeError('Code "{}" not recognized.'.format(code))